Use Page component instead for header customization support
