+++ package.json
@@ -9,10 +9,12 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "bootstrap": "^5.3.3",
     "chart.js": "^4.4.2",
     "js-yaml": "^4.1.0",
     "next": "14.1.3",
     "react": "^18",
+    "react-bootstrap": "^2.10.2",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18",
     "react-tooltip": "^5.26.3"
