--- src/app/header.css
@@ -1,16 +1,0 @@
-/* Header.css */
-
-.header {
-  background-color: var(--card-color);
-  padding: 10px;
-  border-bottom: 2px solid var(--card-border-color);
-  width: 100%;
-  height: 60px;
-  display: grid;
-  grid-template-columns: 100px 1fr 100px;
-  align-items: center;
-}
-
-.pageTitle{
-  margin:auto;
-}
