+++ src/app/instructor/assignments/create/page.tsx
@@ -2,6 +2,7 @@
 
 import React, { useState } from 'react';
 import styles from './page.module.css';
+import Page, { Header, Body } from "../../../../components/Page";
 
 const CreateAssignment = () => {
   const [assignmentName, setAssignmentName] = useState('');
@@ -38,69 +39,74 @@ const CreateAssignment = () => {
   };
 
   return (
-    <div>
-      <h1>Create Assignment</h1>
-      <div className={styles.assignmentForm}>
-        <label htmlFor="assignmentName">Assignment Name:</label>
-        <input
-          type="text"
-          id="assignmentName"
-          value={assignmentName}
-          onChange={(e) => setAssignmentName(e.target.value)}
-        />
-        <label htmlFor="dueDate">Due Date:</label>
-        <input
-          type="date"
-          id="dueDate"
-          value={dueDate}
-          onChange={(e) => setDueDate(e.target.value)}
-        />
-        <label htmlFor="associatedClass">Associated Class:</label>
-        <select
-          id="associatedClass"
-          value={associatedClass}
-          onChange={(e) => setAssociatedClass(e.target.value)}
-        >
-          <option value="">Select Class</option>
-          {/* Populate dropdown with classes */}
-          <option value="CMPUT 101">CMPUT 101</option>
-          <option value="CMPUT 102">CMPUT 102</option>
-          {/* Add more options as needed */}
-        </select>
-        <div className={styles.rubric}>
-          <h3>Rubric</h3>
-          {rubricCategories.map((category, index) => (
-            <div key={index} className={styles.category}>
-              <input
-                type="text"
-                placeholder="Category Name"
-                value={category.name}
-                onChange={(e) => handleCategoryChange(index, 'name', e.target.value)}
-              />
-              <input
-                type="number"
-                min="0"
-                max="100"
-                placeholder="Max Marks"
-                value={category.maxMarks}
-                onChange={(e) => handleCategoryChange(index, 'maxMarks', parseInt(e.target.value))}
-              />
-              <input
-                type="number"
-                min="0"
-                max={category.maxMarks}
-                placeholder="Marks"
-                value={category.marks}
-                onChange={(e) => handleRubricValueChange(index, parseInt(e.target.value))}
-              />
-              <button onClick={() => handleRemoveCategory(index)}>Remove</button>
+    <Page>
+      <Header text="Create Assignment" />
+      <Body>
+        <div>
+          <h1>Create Assignment</h1>
+          <div className={styles.assignmentForm}>
+            <label htmlFor="assignmentName">Assignment Name:</label>
+            <input
+              type="text"
+              id="assignmentName"
+              value={assignmentName}
+              onChange={(e) => setAssignmentName(e.target.value)}
+            />
+            <label htmlFor="dueDate">Due Date:</label>
+            <input
+              type="date"
+              id="dueDate"
+              value={dueDate}
+              onChange={(e) => setDueDate(e.target.value)}
+            />
+            <label htmlFor="associatedClass">Associated Class:</label>
+            <select
+              id="associatedClass"
+              value={associatedClass}
+              onChange={(e) => setAssociatedClass(e.target.value)}
+            >
+              <option value="">Select Class</option>
+              {/* Populate dropdown with classes */}
+              <option value="CMPUT 101">CMPUT 101</option>
+              <option value="CMPUT 102">CMPUT 102</option>
+              {/* Add more options as needed */}
+            </select>
+            <div className={styles.rubric}>
+              <h3>Rubric</h3>
+              {rubricCategories.map((category, index) => (
+                <div key={index} className={styles.category}>
+                  <input
+                    type="text"
+                    placeholder="Category Name"
+                    value={category.name}
+                    onChange={(e) => handleCategoryChange(index, 'name', e.target.value)}
+                  />
+                  <input
+                    type="number"
+                    min="0"
+                    max="100"
+                    placeholder="Max Marks"
+                    value={category.maxMarks}
+                    onChange={(e) => handleCategoryChange(index, 'maxMarks', parseInt(e.target.value))}
+                  />
+                  <input
+                    type="number"
+                    min="0"
+                    max={category.maxMarks}
+                    placeholder="Marks"
+                    value={category.marks}
+                    onChange={(e) => handleRubricValueChange(index, parseInt(e.target.value))}
+                  />
+                  <button onClick={() => handleRemoveCategory(index)}>Remove</button>
+                </div>
+              ))}
+              <button onClick={handleAddCategory}>Add Category</button>
             </div>
-          ))}
-          <button onClick={handleAddCategory}>Add Category</button>
+          </div>
+          <button onClick={handleCreateAssignment}>Create Assignment</button>
         </div>
-      </div>
-      <button onClick={handleCreateAssignment}>Create Assignment</button>
-    </div>
+      </Body>
+    </Page>
   );
 };
