+++ src/app/instructor/assignments/review/page.module.css
@@ -1,7 +1,7 @@
 .main-layout {
   display: flex;
   height: 100vh;
-  max-height: calc(100vh - 60px);
+  max-height: calc(100vh - 70px);
   overflow: auto;
 }
 
@@ -25,7 +25,7 @@
   padding: 10px;
 
   position: fixed;
-  height: calc(100vh - 60px);
+  height: calc(100vh - 70px);
 }
 
 .left-column-content {
@@ -73,7 +73,7 @@
   align-items: stretch;
   flex: 1;
   max-width: 80%;
-  min-height: calc(100vh - 60px);
+  min-height: calc(100vh - 70px);
   height: max-content;
   overflow-y: auto;
   margin-left: 20vw;
