+++ src/app/instructor/assignments/review/page.tsx
@@ -1,90 +1,12 @@
-'use client'
+'use client';
 
 import React, { useEffect, useRef, useState } from 'react';
 import styles from './page.module.css';
 import * as yaml from 'js-yaml'; // Import the js-yaml library
+import Page, { Header, Body } from "../../../../components/Page";
+import InstructorGradeSummary from './InstructorGradeSummary';
 
-const InstructorGradeSummary = () => {
-  // Assuming rubric data and student marks data are available
-  const rubricData = [
-    { category: "Grammar", max_marks: 10 },
-    { category: "Content", max_marks: 20 }
-    // Add more rubric categories as needed
-  ];
-
-  const studentMarksData = useRef([
-    { category: "Grammar", student_marks: 8 },
-    { category: "Content", student_marks: 15 }
-    // Add student marks for each category
-  ]);
-
-  function refreshMarks() {
-    setTotalStudentMarks(Array.from(document.querySelectorAll(`input[data-rubric-grading-input]`))
-      .map(element => parseInt((element as HTMLInputElement).value))
-      .reduce((a, b) => a + b, 0));
-  }
-
-  // Calculate total marks and total student marks
-  const totalMarks = rubricData.reduce((total, { max_marks }) => total + max_marks, 0);
-  const [totalStudentMarks, setTotalStudentMarks] = useState(studentMarksData.current.reduce((total, { student_marks }) => total + student_marks, 0));
-  const percentageMark = ((totalStudentMarks / totalMarks) * 100).toFixed(1);
-
-  // Function to show rubric
-  const handleShowRubric = () => {
-    // Logic to show rubric
-    console.log("Show rubric");
-  };
-
-  const handleCompleteMarking = () => {
-    // Logic to complete marking
-    console.log("Complete marking");
-    location.href = "/instructor/assignments";
-  }
-
-  return (
-    <div className={styles['left-column-content']}>
-      <button className={styles['grade-summary-button']} onClick={handleShowRubric}>
-        Show Rubric
-      </button>
-      <table className={styles['grade-summary-table']}>
-        <thead>
-          <tr>
-            <th>Category</th>
-            <th>Grade</th>
-          </tr>
-        </thead>
-        <tbody>
-          {rubricData.map(({ category, max_marks }, index) => {
-            const studentMarks = studentMarksData.current[index]!.student_marks || 0;
-            return (
-              <tr key={category}>
-                <td>{category}</td>
-                <td>
-                  <input type="number" max={max_marks} min={0} value={studentMarks} onChange={e => {
-                    studentMarksData.current[index].student_marks = parseInt(e.target.value);
-                    refreshMarks();
-                  }} data-rubric-grading-input />
-                  /{max_marks}
-                </td>
-              </tr>
-            );
-          })}
-          <tr className={styles['grade-summary-total']}>
-            <td>Total</td>
-            <td>{`${totalStudentMarks}/${totalMarks} = ${percentageMark}%`}</td>
-          </tr>
-        </tbody>
-      </table>
-      <button onClick={handleCompleteMarking}>
-        Complete Marking
-      </button>
-    </div>
-  );
-};
-
-const InstructorReviewPage = () => {
-  // Example YAML content (replace with actual content loaded from file)
-  const yamlContent = `
+const yamlContent = `
     rubric:
       - category: Grammar
         max_marks: 10
@@ -114,6 +36,9 @@ const InstructorReviewPage = () => {
       - comment: YAAAAAA
   `;
 
+
+
+const InstructorReviewPage = () => {
   // Parse YAML content
   const [data, setData] = useState<{
     rubric: { category: string, max_marks: number }[],
@@ -150,7 +75,7 @@ const InstructorReviewPage = () => {
             const commentDiv = document.querySelector(`.${styles['comment']}[data-comment-id="${id}"]`) as HTMLDivElement;
             commentDiv.style.position = 'absolute';
 
-            let positionedTop = highlightedText.offsetTop - 60; // minus 60 bc the header is 60px tall
+            let positionedTop = highlightedText.offsetTop - 70; // minus 70 bc the header is 70px tall
             if (i > 0) {
               // Check if we are overlapping with predecessor
               if (cannotBeLessThan > positionedTop) {
@@ -186,7 +111,7 @@ const InstructorReviewPage = () => {
         const commentDiv = document.querySelector(`.${styles['comment']}[data-comment-id="${id}"]`) as HTMLDivElement;
         commentDiv.style.position = 'absolute';
 
-        let positionedTop = highlightedText.offsetTop - 60; // minus 60 bc the header is 60px tall
+        let positionedTop = highlightedText.offsetTop - 70; // minus 70 bc the header is 70px tall
         if (i > 0) {
           // Check if we are overlapping with predecessor
           if (cannotBeLessThan > positionedTop) {
@@ -267,54 +192,66 @@ const InstructorReviewPage = () => {
 
   // Render comments with highlighted text
   return (
-    <div className={styles['main-layout']}>
-      <div className={styles['review-page']}>
-        <div className={styles['left-column']}>
-          <h2>English</h2>
-          <div className={`${styles['assignment-details']}`}> {/* Apply left-column-content style here */}
-            <h3>Rubric Details</h3>
-          </div>
-          <div className={`${styles['assignment-details-widget']}`}>
-            {InstructorGradeSummary()}
-          </div>
-        </div>
-        <div className={styles['assignment-and-comments']}>
-          <div className={styles['center-column']}>
-            <div className={styles['assignment-header']}>
-              <h1>Assignment Name</h1>
+    <Page>
+      <Header>
+        <h1>Review</h1>
+        {/* <div>
+          <Form>
+            <Form.Check type="switch" label="AI Review" />
+          </Form>
+        </div> */}
+      </Header>
+      <Body>
+        <div className={styles['main-layout']}>
+          <div className={styles['review-page']}>
+            <div className={styles['left-column']}>
+              <h2>English</h2>
+              <div className={`${styles['assignment-details']}`}> {/* Apply left-column-content style here */}
+                <h3>Rubric Details</h3>
+              </div>
+              <div className={`${styles['assignment-details-widget']}`}>
+                {InstructorGradeSummary()}
+              </div>
             </div>
-            <div className={styles['ipsum-content']}>
-              <h2>Assignment Content</h2>
-              {/* Split the assignment content and wrap sections corresponding to comments with spans */}
-              {comments.map((comment: { id: number, content: string, position: { start: number, end: number } }, index: number) => {
-                const { start, end } = comment.position;
-                const beforeComment = index === 0 ? assignmentContent.slice(0, start) : "";
-                const commentText = assignmentContent.slice(start, end + 1);
+            <div className={styles['assignment-and-comments']}>
+              <div className={styles['center-column']}>
+                <div className={styles['assignment-header']}>
+                  <h1>Assignment Name</h1>
+                </div>
+                <div className={styles['ipsum-content']}>
+                  <h2>Assignment Content</h2>
+                  {/* Split the assignment content and wrap sections corresponding to comments with spans */}
+                  {comments.map((comment: { id: number, content: string, position: { start: number, end: number } }, index: number) => {
+                    const { start, end } = comment.position;
+                    const beforeComment = index === 0 ? assignmentContent.slice(0, start) : "";
+                    const commentText = assignmentContent.slice(start, end + 1);
 
-                const afterCommentEndPos = index === comments.length - 1 ? assignmentContent.length : comments[index + 1].position.start;
-                const afterComment = assignmentContent.slice(end + 1, afterCommentEndPos);
-                return (
-                  <React.Fragment key={index}>
-                    <span>{beforeComment}</span>
-                    <span
-                      className={styles['highlighted-text']}
-                      data-comment-id={comment.id}
-                      onClick={onCommentClick}
-                    >
-                      {commentText}
-                    </span>
-                    <span>{afterComment}</span>
-                  </React.Fragment>
-                );
-              })}
+                    const afterCommentEndPos = index === comments.length - 1 ? assignmentContent.length : comments[index + 1].position.start;
+                    const afterComment = assignmentContent.slice(end + 1, afterCommentEndPos);
+                    return (
+                      <React.Fragment key={index}>
+                        <span>{beforeComment}</span>
+                        <span
+                          className={styles['highlighted-text']}
+                          data-comment-id={comment.id}
+                          onClick={onCommentClick}
+                        >
+                          {commentText}
+                        </span>
+                        <span>{afterComment}</span>
+                      </React.Fragment>
+                    );
+                  })}
+                </div>
+              </div>
+              <div className={styles['right-column']}>
+                {commentDivs}
+              </div>
             </div>
           </div>
-          <div className={styles['right-column']}>
-            {commentDivs}
-          </div>
         </div>
-      </div>
-    </div>
+      </Body>
+    </Page>
   );
 };
