+++ src/components/Page.module.css
@@ -1,0 +1,14 @@
+.header {
+    background-color: var(--card-color);
+    padding: 10px;
+    border-bottom: 2px solid var(--card-border-color);
+    width: 100%;
+    height: 70px;
+    display: grid;
+    grid-template-columns: 150px 1fr 150px;
+    align-items: center;
+}
+
+.header h1 {
+    margin: auto;
+}
